Support for Python3.
Tested on linux (lubuntu 16.04) with Python 3.5 and Python 2.7 using
'ctest' and 'make test_python_psc'.

+++ examples/python/x13.py
@@ -22,10 +22,10 @@
 from numpy import *
 
 text = ["Maurice",
-	"Geoffrey",
-	"Alan",
-	"Rafael",
-	"Vince"]
+        "Geoffrey",
+        "Alan",
+        "Rafael",
+        "Vince"]
 
 # main
 #
@@ -46,34 +46,34 @@ def main(w):
     theta0 = 0
 
     for i in range(5):
-	#theta quantities scaled as integers to make logic independent of
-	#floating point considerations.
-	#n.b. theta multiplied by 2 pi/500 afterward so per is in fact
-	#interpreted as a percentage.
-	theta1 = theta0 + int(5.*per[i])
-	if i == 4:
-	    theta1 = 500
+        #theta quantities scaled as integers to make logic independent of
+        #floating point considerations.
+        #n.b. theta multiplied by 2 pi/500 afterward so per is in fact
+        #interpreted as a percentage.
+        theta1 = theta0 + int(5.*per[i])
+        if i == 4:
+            theta1 = 500
 
-	theta = arange(theta0,theta1+1)
-	x = concatenate((array([5.]), (5. + 3.*cos((2.*pi/500.)*theta))))
-	y = concatenate((array([5.]), (5. + 3.*sin((2.*pi/500.)*theta))))
+        theta = arange(theta0,theta1+1)
+        x = concatenate((array([5.]), (5. + 3.*cos((2.*pi/500.)*theta))))
+        y = concatenate((array([5.]), (5. + 3.*sin((2.*pi/500.)*theta))))
 
-	w.plcol0(i + 1)
-	w.plpsty((i + 3) % 8 + 1)
-	w.plfill(x, y)
-	w.plcol0(1)
-	w.plline(x, y)
-	just = (2.*pi/500.)*(theta0 + theta1) / 2.
-	dx = .25 * cos(just)
-	dy = .25 * sin(just)
-	if (theta0 + theta1) < 250 or (theta0 + theta1) > 750:
-	    just = 0.
-	else:
-	    just = 1.
+        w.plcol0(i + 1)
+        w.plpsty((i + 3) % 8 + 1)
+        w.plfill(x, y)
+        w.plcol0(1)
+        w.plline(x, y)
+        just = (2.*pi/500.)*(theta0 + theta1) / 2.
+        dx = .25 * cos(just)
+        dy = .25 * sin(just)
+        if (theta0 + theta1) < 250 or (theta0 + theta1) > 750:
+            just = 0.
+        else:
+            just = 1.
 
-	mid = len(x) / 2
-	w.plptex((x[mid] + dx), (y[mid] + dy), 1.0, 0.0, just, text[i])
-	theta0 = theta[-1]
+        mid = int(len(x) / 2)
+        w.plptex((x[mid] + dx), (y[mid] + dy), 1.0, 0.0, just, text[i])
+        theta0 = theta[-1]
 
     w.plfont(2)
     w.plschr(0., 1.3)
